AEnemyJump
extended AEnemy as well as setup a secondary move option so when jumping enemy can move differently than normal.
