+++ Assets/!GeometeryWars/Enemies/!Surface/Prefabs/Surface_Wobble.prefab
@@ -78,7 +78,7 @@ MonoBehaviour:
   mapLayer:
     serializedVersion: 2
     m_Bits: 8192
-  body: {fileID: 0}
+  body: {fileID: 7729936663155646675}
   isActive: 0
   timeWait: 1
   speedThrust: 2.5
