+++ Assets/!GeometeryWars/Enemies/!Surface/Scripts/Enemies/AEnemy.cs
@@ -20,6 +20,8 @@ namespace GeometeryWars
         public Transform target;
         protected EnemyMovement currentMovement;
 
+        private Coroutine delay = null;
+
         [Tooltip("The amount of points recieved for destroying this enemy")]
         public int value = 100;
 
@@ -49,7 +51,7 @@ namespace GeometeryWars
             ChildTrigger[] ct = GetComponentsInChildren<ChildTrigger>();
             foreach(ChildTrigger c in ct)
             {
-                Debug.Log("Child Trigger Assigned");
+                //Debug.Log("Child Trigger Assigned");
                 c.TRIGGER += ReturnSelf;
             }
 
@@ -57,7 +59,7 @@ namespace GeometeryWars
             transform.rotation = Quaternion.AngleAxis(UnityEngine.Random.Range(0f, 360f), transform.up) * transform.rotation;
             //start movement delay
             isActive = false;
-            StartCoroutine(Wait());
+            delay = StartCoroutine(CoroutineEX.Delay(() => { isActive = true; }, timeWait));
         }
 
         private void OnDisable()
@@ -101,7 +103,7 @@ namespace GeometeryWars
 
         protected virtual void OnTriggerEnter(Collider other)
         {
-            ReturnToPool(gameObject);
+            ReturnSelf();
         }
 
         private void ReturnSelf()
@@ -112,121 +114,5 @@ namespace GeometeryWars
 
         //METHODS
         protected abstract void SetMovement();
-
-        //COROUTINES
-        //for now a way to prevent enemies from moving right away
-        IEnumerator Wait()
-        {
-            yield return new WaitForSeconds(timeWait);
-            isActive = true;
-        }
-    }
-
-
-    //---------------------------------------------------
-    public abstract class EnemyMovement
-    {
-        public EnemyMovement(AEnemy e)
-        {
-            this.e = e;
-        }
-
-        protected AEnemy e;
-
-        public virtual Vector3 RayDirection()
-        {
-            return e.transform.forward * e.speedThrust * Time.fixedDeltaTime;
-        }
-        public virtual Vector3 NextPosition(RaycastHit hit)
-        {
-            return hit.point + hit.normal;
-        }
-        public virtual Quaternion NextRotation(RaycastHit hit)
-        {
-            return Quaternion.FromToRotation(e.transform.up, hit.normal) * e.transform.rotation;
-        }
-    }
-
-    //basic movement, no rotation just move forward
-    public class wander : EnemyMovement
-    {
-        public wander(AEnemy e) : base(e)
-        {
-        }
-    }
-
-    //rotate to face target
-    public class RotateToTarget : EnemyMovement
-    {
-        public RotateToTarget(AEnemy e) : base(e)
-        {
-        }
-
-        public override Vector3 RayDirection()
-        {
-            return Vector3.zero;
-        }
-
-        public override Quaternion NextRotation(RaycastHit hit)
-        {
-            Vector3 selfToTarget = (e.target.position - e.transform.position).normalized;
-            e.transform.rotation = Quaternion.FromToRotation(e.transform.forward, selfToTarget) * e.transform.rotation;
-            e.transform.rotation = Quaternion.FromToRotation(e.transform.up, hit.normal) * e.transform.rotation;
-            return e.transform.rotation;
-        }
-    }
-
-    //rotate to face, move in faced direction
-    public class Chase : EnemyMovement
-    {
-        public Chase(AEnemy e) : base(e)
-        {
-        }
-
-        public override Quaternion NextRotation(RaycastHit hit)
-        {
-            Vector3 selfToTarget = (e.target.position - e.transform.position).normalized;
-            e.transform.rotation = Quaternion.FromToRotation(e.transform.forward, selfToTarget) * e.transform.rotation;
-            e.transform.rotation = Quaternion.FromToRotation(e.transform.up, hit.normal) * e.transform.rotation;
-            return e.transform.rotation;
-        }
-    }
-
-    //chase but with horizontal offset in movement
-    public class Wobble : EnemyMovement
-    {
-        public Wobble(AEnemy e, Vector3 wobbleDirection, float wobbleSpeed, float wobbleStrength) : base(e)
-        {
-            this.wobbleDirection = wobbleDirection;
-            this.wobbleSpeed = wobbleSpeed;
-            this.wobbleStrength = wobbleStrength;
-
-            //generate random seed time so not all wobbles look exactly alike
-            seedTime = UnityEngine.Random.Range(0f, DateTime.Now.Millisecond);
-        }
-
-        private Vector3 wobbleDirection;
-        private float wobbleSpeed, wobbleStrength;
-
-        private float seedTime = 0f;
-
-        //incorporate wobble offset
-        public override Vector3 RayDirection()
-        {
-            Vector3 wobble = e.transform.TransformDirection(wobbleDirection) * Mathf.Sin((seedTime + Time.time) * wobbleSpeed) * wobbleStrength;
-            //TEST FOR OTHER WOBBLE TYPES
-            //Vector3 wobble2 = e.transform.TransformDirection(Vector3.forward) * Mathf.Sin((seedTime + Time.time) * (wobbleSpeed * 2)) * (wobbleStrength * 0.75f);
-            //Vector3 total = (wobble + wobble2);
-            return (wobble + e.transform.forward) * e.speedThrust * Time.fixedDeltaTime;
-        }
-
-        //rotate to face target
-        public override Quaternion NextRotation(RaycastHit hit)
-        {
-            Vector3 selfToTarget = (e.target.position - e.transform.position).normalized;
-            e.transform.rotation = Quaternion.FromToRotation(e.transform.forward, selfToTarget) * e.transform.rotation;
-            e.transform.rotation = Quaternion.FromToRotation(e.transform.up, hit.normal) * e.transform.rotation;
-            return e.transform.rotation;
-        }
     }
 }
