+++ Assets/!GeometeryWars/Enemies/!Surface/Scripts/Enemies/AEnemyJump.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bac2bd8a60110ac4c94fc6e8091dc00d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
