+++ Assets/!GeometeryWars/Enemies/!Surface/Scripts/Enemies/EnemyMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 90b639c49dbd2b340ad877995a08363f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
