+++ Assets/!GeometeryWars/Enemies/!Surface/Scripts/Other/CoroutineEX.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e71db593de8a79447a70eb124c0e1cb7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
