+++ Assets/!GeometeryWars/Spawning/TestLevel/Level_01.asset
@@ -28,7 +28,7 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 7e50339b9d60ef24cb8ce42fd188f101, type: 2}
   - {fileID: 11400000, guid: 9d68d3aadd199c7489a70a1cc9fa8e49, type: 2}
   arraySize: 6
-  enemyTypeIndex: 030000000400000002000000010000000000000005000000
+  enemyTypeIndex: 010000000400000001000000040000000100000004000000
   patternIndex: 040000000100000002000000030000000000000000000000
   spawnTimes:
   - 3
